Merge in ONE/products-exp-service from hotfix/fixed-issue to release/2.5.0-rc

* commit '11d903456c270af8a0f94bcf9cd59ca616348b2d':
  fixed issue calculate range income invalid

+++ src/main/java/com/tmb/oneapp/productsexpservice/model/loan/RangeIncome.java
@@ -5,6 +5,8 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
@@ -16,5 +18,5 @@ public class RangeIncome {
     private int minAmount;
     private int maxAmount;
     private String statusWorking;
-    private long revenueMultiple;
+    private BigDecimal revenueMultiple;
 }
